test that you can auto-update from the just-packaged version
